requirements.txt
discord.py==2.3.2
gspread==6.0.0
oauth2client==4.1.3
python-dotenv==1.0.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0
pandas==2.1.4
pytz==2024.1
requests==2.31.0
